Implemented reload speed mixin.

+++ lua/Factions/Weapons/Factions_DamageMixin.lua
@@ -32,7 +32,7 @@ DamageMixin.expectedConstants =
 
 DamageMixin.networkVars =
 {
-	damageScalar = "decimal",
+	damageScalar = "float",
 }
 
 function DamageMixin:__initmixin()
