+++ lua/Factions/Weapons/Factions_FireRateMixin.lua
@@ -32,7 +32,7 @@ FireRateMixin.expectedConstants =
 
 FireRateMixin.networkVars =
 {
-	fireRateScalar = "decimal",
+	fireRateScalar = "float",
 }
 
 function FireRateMixin:__initmixin()
